# weather_app/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WeatherApi - OpenAPI 3.0",
    "description": "This coding exercise will help us understand how you approach some of the common problems we see in data engineering. Ask questions if things are unclear, use best practices and common software patterns, and feel free to go the extra mile to show off your skills. Imagine you are handing off your completed project to someone else to maintain -- it should be clear to another developer how things work and your reasoning behind your design decisions.",
    "version": "1.0.11"
  },
  "tags": [
    {
      "name": "Weather Data",
      "description": "max_temp , min_temp and precipitatioin data."
    },
    {
      "name": "Weather Statistics",
      "description": "Mean of weather data of each year."
    }
  ],
  "paths": {
    "/api/weather": {
      "get": {
        "tags": [
          "Weather Data"
        ],
        "summary": "Weather data of a few stations.",
        "description": "Returns weather data of multiple weather stations around USA with with the time they are recorded.",
        "operationId": "getWeatherData",
        "parameters": [
          {
            "name": "station_code",
            "in": "query",
            "description": "This is a unique station code.",
            "required": false,
            "explode": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "query",
            "description": "This is the date of when the weather data was recorded IN YYYY-MM-DD format.",
            "required": false,
            "explode": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page_no",
            "in": "query",
            "description": "Page number of response.",
            "required": false,
            "explode": true,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Page size of response.",
            "required": false,
            "explode": true,
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "payload": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Weather_Data"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Data not Found!"
          }
        }
      }
    },
    "/api/weather/stats": {
      "get": {
        "tags": [
          "Weather Statistics"
        ],
        "summary": "Statistics of Weather around US.",
        "description": "Returns average Weather Statistics of each year collected from weather stations around the USA.",
        "operationId": "weatherStats",
        "parameters": [
          {
            "name": "station_code",
            "in": "query",
            "description": "This is a unique station code.",
            "required": false,
            "explode": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "in": "query",
            "description": "This is the year of the.",
            "required": false,
            "explode": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page_no",
            "in": "query",
            "description": "Page number of response.",
            "required": false,
            "explode": true,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Page size of response.",
            "required": false,
            "explode": true,
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "payload": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Weather_Stats"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Data not Found!"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Weather_Data": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "station_code": {
            "type": "string",
            "example": "USC11001293"
          },
          "date": {
            "type": "string",
            "example": "1985-01-01"
          },
          "max_temp": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "min_temp": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "prcp": {
            "type": "integer",
            "format": "int64",
            "example": 1
          }
        },
        "xml": {
          "name": "Weather_Data"
        }
      },
      "Weather_Stats": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "year": {
            "type": "integer",
            "format": "int64",
            "example": 1985
          },
          "station_code": {
            "type": "string",
            "example": "USC11001293"
          },
          "avg_max_temp": {
            "type": "number",
            "format": "float",
            "example": 3.14
          },
          "avg_min_temp": {
            "type": "number",
            "format": "float",
            "example": 3.14
          },
          "total_acc_prcp": {
            "type": "integer",
            "format": "int64",
            "example": 1
          }
        }
      }
    }
  }
}